Split out web6 packages to functional files
This helps when debugging, instead of working in one big file you can
focus on what you need to.
